+++ tsconfig.json
@@ -13,6 +13,6 @@
         }
     },
     "exclude" :[
-        "node_modules"
+        "node_modules", "src/components/SubSection"
     ],
 }
